+++ src/coreclr/vm/ecalllist.h
@@ -687,9 +687,9 @@ FCFuncStart(gWeakReferenceOfTFuncs)
     FCFuncElement("IsTrackResurrection", WeakReferenceOfTNative::IsTrackResurrection)
 FCFuncEnd()
 
-FCFuncStart(gInteropComWeakReferenceHelpers)
-    FCFuncElement("ComWeakRefToObject", WeakReferenceNative::ComWeakRefToObject)
-    FCFuncElement("ObjectToComWeakRef", WeakReferenceNative::ObjectToComWeakRef)
+FCFuncStart(gComAwareWeakReferenceFuncs)
+    FCFuncElement("ComWeakRefToObject", ComAwareWeakReferenceNative::ComWeakRefToObject)
+    FCFuncElement("ObjectToComWeakRef", ComAwareWeakReferenceNative::ObjectToComWeakRef)
 FCFuncEnd()
 
 #ifdef FEATURE_COMINTEROP
@@ -729,7 +729,7 @@ FCClassElement("Array", "System", gArrayFuncs)
 FCClassElement("AssemblyLoadContext", "System.Runtime.Loader", gAssemblyLoadContextFuncs)
 FCClassElement("Buffer", "System", gBufferFuncs)
 FCClassElement("CastHelpers", "System.Runtime.CompilerServices", gCastHelpers)
-FCClassElement("ComWeakReferenceHelpers", "System.Runtime.InteropServices", gInteropComWeakReferenceHelpers)
+FCClassElement("ComAwareWeakReference", "System", gComAwareWeakReferenceFuncs)
 FCClassElement("CompatibilitySwitch", "System.Runtime.Versioning", gCompatibilitySwitchFuncs)
 FCClassElement("CustomAttribute", "System.Reflection", gCOMCustomAttributeFuncs)
 FCClassElement("CustomAttributeEncodedArgument", "System.Reflection", gCustomAttributeEncodedArgument)
